requirements.txt
# FILE: requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.4
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangChain and LangGraph
# langchain==0.1.7
# langchain-community==0.0.20
# langchain-core==0.1.23
# langgraph==0.0.20

# LangChain and LangGraph
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37  # <-- FINAL REQUIRED CHANGE
langgraph==0.1.1

# LLM Providers
openai==1.12.0
anthropic==0.18.1
huggingface-hub==0.20.3
google-genai

# Vector store and embeddings
# faiss-cpu==1.7.4
faiss-cpu==1.12.0
sentence-transformers==2.3.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Image generation
diffusers==0.25.1
# torch==2.1.2
torch==2.9.1
transformers==4.37.2

# Voice I/O
SpeechRecognition==3.10.1

# sx3==2.90
pyttsx3==2.90
vosk==0.3.45

# Diagram generation
graphviz==0.20.1

# Streamlit
streamlit==1.32.0

# Data and utilities
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3
aiofiles==23.2.1

# Testing
# pytest==8.0.0
pytest==8.2.0
# pytest-asyncio==0.23.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

python-multipart

scipy
pytesseract

kokoro
soundfile